Make console output lines stream indefinitely
Updated the code that streams console lines to serve them indefinitely.
This makes development of a frontend way easier, as you no longer need
to manually check if a line is new/unique, it will send it when they
are added to the scrollback buffer periodically.

Need to test if this results in timeouts after extended periods of time,
only ran into this because I'm working on an example frontend to make
certain its not like pulling teeth writing code for this