Added player analytics and rewrote db classes
Added code that tracks player sessions to your servers for future
use in an endpoint or two

Re-wrote the database class to be more generic to allow for easier
extension for future and current features. Also added some classes for
the API database and the server/player analytics databases.

This lets me add code to the database classes that is specific to a
certain use case without needing to crowd the existing class or re-write
it in its entirety.