Add UUID cache invalidation
In order to ensure that the UUID lookups are correct for a given user, I
unfortunately need to force a lookup for every new session. This ensures
the username we use is always owned but the correct UUID we get.

This drastically lowers cache effectiveness (whereas before it was
persistent forever which is actually a bad thing), but will ensure the
code remains as safe as possible