Add timestamps to console output
Added timestamps to the console output to make it easier to determine if
a new line is unique or not when making a frontend.
This was done to fix some issues I ran into developing a frontend for
this project, essentially I had issues populating the console window
while allowing for switching servers.

Note: The strucure of the console output endpoint may be changed at
some point to no longer use a streaming response, it seems to cause some
issues when working with python requests for example where your request
cannot be interrupted FOREVER. which is not great.